code with debugging information

+++ src/CellComp.h
@@ -7,23 +7,23 @@ class CellComp : public CellBase {
     Q_OBJECT
     Q_PROPERTY(bool shot READ getShot WRITE setShot NOTIFY shotChanged)
     Q_PROPERTY(bool possibleShot READ getPossibleShot WRITE setPossibleShot NOTIFY possibleShotChanged)
-    Q_PROPERTY(int chance READ getChance WRITE setChance NOTIFY chanceChanged)
+    Q_PROPERTY(double chance READ getChance WRITE setChance NOTIFY chanceChanged)
 
 public:
     explicit CellComp(QObject *parent = 0);
 
     bool getShot();
     bool getPossibleShot();
-    int getChance();
+    double getChance();
 
     void setShot(bool shot);
     void setPossibleShot(bool possibleShot);
-    void setChance(int chance);
+    void setChance(double chance);
 
 private:
     bool shot;
     bool possibleShot;
-    int chance;
+    double chance;
 
 signals:
     void shotChanged();
